+++ src/Content/Presentation/TranslationGroupPresenter.php
@@ -10,7 +10,7 @@ use Kumwe\App\Content\Application\ContentService;
 use Kumwe\App\Content\Application\TranslationGroupRepository;
 use Kumwe\App\Content\Domain\TranslationGroup;
 use Kumwe\App\Content\Domain\TranslationGroupMember;
-use Kumwe\App\Localization\Application\ActiveLocale;
+use Kumwe\Localization\Application\ActiveLocale;
 use Kumwe\App\Site\Application\PublicPageLocator;
 use Locale;
 use Psr\Clock\ClockInterface;