+++ tests/Support/TransientBusinessDefinitionFixtureScope.php
@@ -6,7 +6,7 @@ namespace Kumwe\App\Tests\Support;
 
 use Doctrine\DBAL\Connection;
 use Kumwe\App\Application\Authorization\SiteContext;
-use Kumwe\App\Application\Persistence\TransactionManager;
+use Kumwe\Transaction\Contract\TransactionManager;
 use Kumwe\App\BusinessDefinition\Application\BusinessDefinitionRepository;
 use Kumwe\App\BusinessDefinition\Domain\DefinitionStatus;
 use Kumwe\App\BusinessSchema\Application\BusinessSchemaInstallationRepository;